+++ IzvorniKod/src/main/resources/application.properties
@@ -24,7 +24,7 @@ spring.jpa.properties.hibernate.use_sql_comments=true
 spring.security.user.name=userN
 spring.security.user.password=passwordN
 #server.port=8080
-
+#server.ssl.enabled=false
 server.port=8443
 #server.address=0.0.0.0
 server.ssl.key-store=classpath:keystore.p12
